--- src/main/java/com/sparta/threello/Board.java
@@ -1,4 +1,0 @@
-package com.sparta.threello;
-
-public class Board {
-}
