--- src/main/java/com/sparta/threello/Deck.java
@@ -1,4 +1,0 @@
-package com.sparta.threello;
-
-public class Deck {
-}
