+++ src/main/java/com/sparta/threello/entity/CardDetail.java
@@ -1,0 +1,4 @@
+package com.sparta.threello.entity;
+
+public class CardDetail {
+}
