+++ src/main/java/com/sparta/threello/entity/Deck.java
@@ -1,0 +1,4 @@
+package com.sparta.threello.entity;
+
+public class Deck {
+}
